+++ tools/integration_tests/managed_folders/empty_managed_folder_list_test.go
@@ -12,10 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package empty_managed_folders_list
+// Provides integration tests for listing empty managed folders.
+package managed_folders
 
 import (
 	"fmt"
+	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/operations"
+	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/test_setup"
 	"io/fs"
 	"log"
 	"os"
@@ -23,15 +26,39 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/client"
-	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/operations"
+	"github.com/googlecloudplatform/gcsfuse/internal/config"
+
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup"
 )
+
+const (
+	testDirName                     = "EmptyManagedFoldersTest"
+	NumberOfObjectsInDirForListTest = 4
+	EmptyManagedFolder1             = "emptyManagedFolder1"
+	EmptyManagedFolder2             = "emptyManagedFolder2"
+	SimulatedFolder                 = "simulatedFolder"
+	File                            = "testFile"
+)
+
+////////////////////////////////////////////////////////////////////////
+// Boilerplate
+////////////////////////////////////////////////////////////////////////
+
+type enableEmptyManagedFoldersTrue struct {
+	flags []string
+}
+
+func (s *enableEmptyManagedFoldersTrue) Setup(t *testing.T) {
+	setup.SetupTestDirectory(testDirName)
+}
+
+func (s *enableEmptyManagedFoldersTrue) Teardown(t *testing.T) {
+}
 
 func createDirectoryStructureForTest(t *testing.T) {
 	bucket := setup.TestBucket()
 	testDir := testDirName
-	client.SetBucketAndObjectBasedOnTypeOfMount(&bucket, &testDir)
+	setup.SetBucketAndObjectBasedOnTypeOfMount(&bucket, &testDir)
 
 	operations.CreateManagedFoldersInTestDir(EmptyManagedFolder1, bucket, testDir)
 	operations.CreateManagedFoldersInTestDir(EmptyManagedFolder2, bucket, testDir)
@@ -40,9 +67,7 @@ func createDirectoryStructureForTest(t *testing.T) {
 	operations.CloseFile(f)
 }
 
-func TestListDirectoryForEmptyManagedFolders(t *testing.T) {
-	setup.SetupTestDirectory(testDirName)
-
+func (s *enableEmptyManagedFoldersTrue) TTestListDirectoryForEmptyManagedFolders(t *testing.T) {
 	// Create directory structure for testing.
 	createDirectoryStructureForTest(t)
 
@@ -62,12 +87,12 @@ func TestListDirectoryForEmptyManagedFolders(t *testing.T) {
 		if err != nil {
 			log.Fatal(err)
 		}
-
 		// Check if managedFolderTest directory has correct data.
-		if dir.IsDir() && dir.Name() == testDirName {
+		if dir.Name() == testDirName {
+			fmt.Println(dir.Name())
 			// numberOfObjects - 4
 			if len(objs) != NumberOfObjectsInDirForListTest {
-				t.Errorf("Incorrect number of objects in the directory expectected %d: got %d: ", NumberOfObjectsInDirForListTest, len(objs))
+				t.Errorf("Incorrect number of objects in the directory %s expectected %d: got %d: ", dir.Name(), NumberOfObjectsInDirForListTest, len(objs))
 			}
 
 			// testBucket/managedFolderTest/emptyManagedFolder1   -- ManagedFolder1
@@ -89,15 +114,13 @@ func TestListDirectoryForEmptyManagedFolders(t *testing.T) {
 			if objs[3].Name() != File || objs[3].IsDir() != false {
 				t.Errorf("Listed incorrect object expectected %s: got %s: ", File, objs[3].Name())
 			}
-
 			return nil
 		}
-
 		// Check if subDirectory is empty.
-		if dir.IsDir() && (dir.Name() == EmptyManagedFolder1 || dir.Name() == EmptyManagedFolder2 || dir.Name() == SimulatedFolder) {
+		if dir.Name() == EmptyManagedFolder1 || dir.Name() == EmptyManagedFolder2 || dir.Name() == SimulatedFolder {
 			// numberOfObjects - 0
 			if len(objs) != 0 {
-				t.Errorf("Incorrect number of objects in the directory expectected %d: got %d: ", 0, len(objs))
+				t.Errorf("Incorrect number of objects in the directory %s expectected %d: got %d: ", dir.Name(), 0, len(objs))
 			}
 		}
 
@@ -106,5 +129,53 @@ func TestListDirectoryForEmptyManagedFolders(t *testing.T) {
 	if err != nil {
 		t.Errorf("error walking the path : %v\n", err)
 		return
+	}
+}
+
+func getMountConfigForEmptyManagedFolders() config.MountConfig {
+	mountConfig := config.MountConfig{
+		ListConfig: config.ListConfig{
+			EnableEmptyManagedFolders: true,
+		},
+		LogConfig: config.LogConfig{
+			Severity:        config.TRACE,
+			LogRotateConfig: config.DefaultLogRotateConfig(),
+		},
+	}
+
+	fmt.Println(mountConfig)
+	return mountConfig
+}
+
+////////////////////////////////////////////////////////////////////////
+// TestMain
+////////////////////////////////////////////////////////////////////////
+
+func TestEnableEmptyManagedFoldersTrue(t *testing.T) {
+	ts := &enableEmptyManagedFoldersTrue{}
+
+	setup.ExitWithFailureIfBothTestBucketAndMountedDirectoryFlagsAreNotSet()
+
+	// Run tests for mountedDirectory only if --mountedDirectory  and --testBucket flag is set.
+	if setup.AreBothMountedDirectoryAndTestBucketFlagsSet() {
+		test_setup.RunTests(t, ts)
+		return
+	}
+
+	// Else run tests for testBucket.
+	// Set up test directory.
+	setup.SetUpTestDirForTestBucketFlag()
+
+	configFile := setup.YAMLConfigFile(
+		getMountConfigForEmptyManagedFolders(),
+		"config.yaml")
+
+	flagSet := [][]string{{"--implicit-dirs", "--config-file=" + configFile}}
+
+	// Run tests.
+	for _, flags := range flagSet {
+		ts.flags = flags
+		log.Printf("Running tests with flags: %s", ts.flags)
+		test_setup.RunTests(t, ts)
 	}
 }
