+++ tools/integration_tests/managed_folders/folder_view_permissions_test.go
